EDit Auth System

+++ app/Http/Controllers/Api/ClintLoginController.php
@@ -106,24 +106,29 @@ class ClintLoginController extends Controller
         ]);
 
         $user = User::where('phone', $request->phone)->first();
-        if ($user->type == 1) {
-            $user_id = User::where('phone', $request->phone)->value('id');
+        if ($user != null) {
+            if ($user->type == 1) {
+                $user_id = User::where('phone', $request->phone)->value('id');
 
-            $token = Token::where('tokenable_id', $user_id)->count();
-            if ($token != 0) {
-                $user->tokens()->delete();
+                $token = Token::where('tokenable_id', $user_id)->count();
+                if ($token != 0) {
+                    $user->tokens()->delete();
+                }
+
+                if (!$user || !Hash::check($request->password, $user->password)) {
+                    throw ValidationException::withMessages([
+                        'phone' => ['The provided phone number are incorrect.'],
+                    ]);
+                }
+
+                $token = $user->createToken($request->device_name)->plainTextToken;
+                return response()->json(['token' => $token, 'Data' => $user], 200);
+            } else {
+                return response()->json(['Stutus' => false, 'Message' => 'This App Only For Clint'], 200);
+
             }
-
-            if (!$user || !Hash::check($request->password, $user->password)) {
-                throw ValidationException::withMessages([
-                    'phone' => ['The provided phone number are incorrect.'],
-                ]);
-            }
-
-            $token = $user->createToken($request->device_name)->plainTextToken;
-            return response()->json(['token' => $token, 'Data' => $user], 200);
         } else {
-            return response()->json(['Stutus' => false, 'Message' => 'This App Only For Clint'], 200);
+            return response()->json(['Stutus' => false, 'Message' => 'Your Number Phone Is Wrong'], 200);
 
         }
 
@@ -137,21 +142,25 @@ class ClintLoginController extends Controller
             'device_name' => 'required',
         ]);
         $user = User::where('phone', $request->phone)->first();
-        if ($user->type == 2) {
-            $user_id = User::where('phone', $request->phone)->value('id');
-            $token = Token::where('tokenable_id', $user_id)->count();
-            if ($token != 0) {
-                $user->tokens()->delete();
+        if ($user != null) {
+            if ($user->type == 2) {
+                $user_id = User::where('phone', $request->phone)->value('id');
+                $token = Token::where('tokenable_id', $user_id)->count();
+                if ($token != 0) {
+                    $user->tokens()->delete();
+                }
+                if (!$user || !Hash::check($request->password, $user->password)) {
+                    throw ValidationException::withMessages([
+                        'phone' => ['The provided phone number are incorrect.'],
+                    ]);
+                }
+                $token = $user->createToken($request->device_name)->plainTextToken;
+                return response()->json(['token' => $token, 'Data' => $user], 200);
+            } else {
+                return response()->json(['Stutus' => false, 'Message' => 'This App Only For Seller'], 200);
             }
-            if (!$user || !Hash::check($request->password, $user->password)) {
-                throw ValidationException::withMessages([
-                    'phone' => ['The provided phone number are incorrect.'],
-                ]);
-            }
-            $token = $user->createToken($request->device_name)->plainTextToken;
-            return response()->json(['token' => $token, 'Data' => $user], 200);
         } else {
-            return response()->json(['Stutus' => false, 'Message' => 'This App Only For Seller'], 200);
+            return response()->json(['Stutus' => false, 'Message' => 'Your Number Phone Is Wrong'], 200);
 
         }
 
@@ -166,24 +175,30 @@ class ClintLoginController extends Controller
         ]);
 
         $user = User::where('phone', $request->phone)->first();
-        if ($user->type == 3) {
-            $user_id = User::where('phone', $request->phone)->value('id');
+        $user = User::where('phone', $request->phone)->first();
+        if ($user != null) {
+            if ($user->type == 3) {
+                $user_id = User::where('phone', $request->phone)->value('id');
 
-            $token = Token::where('tokenable_id', $user_id)->count();
-            if ($token != 0) {
-                $user->tokens()->delete();
+                $token = Token::where('tokenable_id', $user_id)->count();
+                if ($token != 0) {
+                    $user->tokens()->delete();
+                }
+
+                if (!$user || !Hash::check($request->password, $user->password)) {
+                    throw ValidationException::withMessages([
+                        'phone' => ['The provided phone number are incorrect.'],
+                    ]);
+                }
+
+                $token = $user->createToken($request->device_name)->plainTextToken;
+                return response()->json(['token' => $token, 'Data' => $user], 200);
+            } else {
+                return response()->json(['Stutus' => false, 'Message' => 'This App Only For Delivery'], 200);
+
             }
-
-            if (!$user || !Hash::check($request->password, $user->password)) {
-                throw ValidationException::withMessages([
-                    'phone' => ['The provided phone number are incorrect.'],
-                ]);
-            }
-
-            $token = $user->createToken($request->device_name)->plainTextToken;
-            return response()->json(['token' => $token, 'Data' => $user], 200);
         } else {
-            return response()->json(['Stutus' => false, 'Message' => 'This App Only For Delivery'], 200);
+            return response()->json(['Stutus' => false, 'Message' => 'Your Number Phone Is Wrong'], 200);
 
         }
